CONTINUE, SKIP_SUBTREE, SKIP_SIBLING, TERMINATE support in S3Walker
